scraper/requirements.txt
# Requires Python 3.11 — run with: py -3.11 -m pip install -r requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
httpx==0.27.0
apscheduler==3.10.4
asyncpg==0.29.0
redis[asyncio]==5.0.4
python-dotenv==1.0.1
